Refactored createPlaylist using mutation helper
