+++ src/modules/playlist/components/PlaylistContainer.jsx
@@ -1,41 +1,14 @@
 import React from 'react';
 
-import {Query, Mutation} from 'react-apollo';
+import {Query} from 'react-apollo';
 import gql from 'graphql-tag';
 
 import {nullToUndefined} from '../../common/utils';
+import CreatePlaylist, {createPlaylistUpdate} from '../mutations/CreatePlaylist';
 import Playlist from './Playlist';
-import PlaylistFragments from '../../common/fragments/PlaylistFragments';
+import ToggleSearch from '../../common/mutations/ToggleSearch';
 import WithPlaylistId from '../../common/components/WithPlaylistId';
 import adapt from '../../common/components/Adapt';
-
-const TOGGLE_SEARCH = gql`
-	mutation ToggleSearch {
-		toggleSearch @client
-	}
-`;
-
-const CREATE_PLAYLIST = gql`
-	mutation CreatePlaylist($playlist: String!) {
-		upsertPlaylist(where: {name: $playlist}, create: {name: $playlist}, update: {}) {
-			...AllPlaylist
-		}
-	}
-
-	${PlaylistFragments.all}
-`;
-
-const createPlaylistUpdate = playlist => (cache, {data: {upsertPlaylist}}) => {
-	const query = gql`
-		query CreatePlaylistUpdate($playlist: String!) {
-			playlist(where: {name: $playlist}) {
-				...AllPlaylist
-			}
-		}
-		${PlaylistFragments.all}
-	`;
-	cache.writeQuery({query, data: {playlist: upsertPlaylist}, variables: {playlist}});
-};
 
 const query = gql`
 	query PlaylistQuery($playlist: String!) {
@@ -54,7 +27,7 @@ const query = gql`
 
 const Composed = adapt(
 	{
-		toggleSearch: <Mutation mutation={TOGGLE_SEARCH} />,
+		toggleSearch: <ToggleSearch nullary />,
 		playlist: <WithPlaylistId />
 	},
 	{
@@ -64,13 +37,9 @@ const Composed = adapt(
 			</Query>
 		),
 		createPlaylist: ({render, playlist}) => (
-			<Mutation
-				mutation={CREATE_PLAYLIST}
-				variables={{playlist}}
-				update={createPlaylistUpdate(playlist)}
-			>
+			<CreatePlaylist variables={{playlist}} update={createPlaylistUpdate(playlist)}>
 				{render}
-			</Mutation>
+			</CreatePlaylist>
 		)
 	}
 );
